Add private DNS server for Tailscale IPs
The homelab now runs a private DNS server. This DNS server has A
records that point to Tailscale IP addresses. This DNS server has been
added to Tailscale's admin console, so that, when connected to Tailscale,
it is used for DNS (see Tailscale's docs https://tailscale.com/kb/1054/dns).
This allows one to use the same domain names as usual when accessing the
homelab remotely through Tailscale.