Add distributed block storage with longhorn
Longhorn lets me decouple storage from pods, so storage does not need to be on the
same node as a pod, as with k3s' default local-path provisioner. We create an
unreplicated storage class to use as the default instead of local-path. Ideally we
would replicate the data, but the homelab is small right now and I don't need
replication for everything.